+++ go/components/spark/job/client/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     deps = [
         "//go/base/env:go_default_library",
         "//go/components/spark/job:go_default_library",
-        "//go/components/spark/job/k8s-crds/apis/sparkoperator.k8s.io/v1beta2:go_default_library",
+        "//go/thirdparty/k8s-crds/apis/sparkoperator.k8s.io/v1beta2:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_go_logr_logr//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
