--- go/components/spark/job/k8s-crds/apis/sparkoperator.k8s.io/README.md
@@ -1,1 +1,0 @@
-To be kept in sync with https://code.uberinternal.com/diffusion/DASPAOV/browse/master/pkg/apis/sparkoperator.k8s.io/
